+++ Runtime/ValueButtonsAttribute.cs
@@ -9,15 +9,20 @@ namespace SaintsField
     public class ValueButtonsAttribute: PathedDropdownAttribute
     {
         public bool NoFold;
+        public EObsolete Obsolete;
 
-        public ValueButtonsAttribute(string funcName = null, EUnique unique = EUnique.None, bool noFold=false): base(funcName, unique)
+        public ValueButtonsAttribute(string funcName = null, EUnique unique = EUnique.None, bool noFold=false,
+            EObsolete obsolete = EObsolete.Remove): base(funcName, unique)
         {
             NoFold = noFold;
+            Obsolete = obsolete;
             slashAsSub = false;
         }
 
         public ValueButtonsAttribute(EUnique unique) : this(null, unique) {}
         public ValueButtonsAttribute(string funcName) : this(funcName, EUnique.None) {}
         public ValueButtonsAttribute(bool noFold) : this(null, EUnique.None, noFold) {}
+        public ValueButtonsAttribute(EObsolete obsolete) : this(null, EUnique.None, false, obsolete) {}
+        public ValueButtonsAttribute(EObsolete obsolete, bool noFold) : this(null, EUnique.None, noFold, obsolete) {}
     }
 }